[query] Add remove query #391 (#392)
* rework queries

* Update query.rs

+++ src/agdb/collections/dictionary/dictionary_data_storage.rs
@@ -111,7 +111,7 @@ where
     Data: Storage,
 {
     fn capacity(&self) -> u64 {
-        self.counts.len() as u64
+        self.counts.len()
     }
 
     fn commit(&mut self) -> Result<(), DbError> {
